// Views/src/components/out-sourced/collect-deliver/components/list/dto/list-collect-deliver-dto.ts
import {ItemsInterface} from '../../../../../../shared/components/list-g/list/interfaces/items-interface'


export class ListCollectDeliverDto {
  id: ItemsInterface;
  destiny: ItemsInterface;
  billingFrom: ItemsInterface;
  start: ItemsInterface;
  price: ItemsInterface;
  subject: ItemsInterface;
  collect: ItemsInterface;
  deliver: ItemsInterface;
  other: ItemsInterface;
  wasPaid: ItemsInterface;
  expires: ItemsInterface;
  expiresView:ItemsInterface;
}